+++ internal/controller/claw_proxy_test.go
@@ -531,149 +531,6 @@ func TestConfigureProxyImage(t *testing.T) {
 	})
 }
 
-func TestResolveProviderDefaults(t *testing.T) {
-	tests := []struct {
-		name       string
-		cred       clawv1alpha1.CredentialSpec
-		wantDomain string
-		wantHeader string
-		wantErr    string
-	}{
-		{
-			name: "google apiKey fills domain and header",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "gemini",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "google",
-			},
-			wantDomain: "generativelanguage.googleapis.com",
-			wantHeader: "x-goog-api-key",
-		},
-		{
-			name: "anthropic apiKey fills domain and header",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "anthropic",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "anthropic",
-			},
-			wantDomain: "api.anthropic.com",
-			wantHeader: "x-api-key",
-		},
-		{
-			name: "google gcp fills domain",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "gemini",
-				Type:     clawv1alpha1.CredentialTypeGCP,
-				Provider: "google",
-				GCP:      &clawv1alpha1.GCPConfig{Project: "p", Location: "us-central1"},
-			},
-			wantDomain: ".googleapis.com",
-		},
-		{
-			name: "anthropic gcp fills domain",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "anthropic-vertex",
-				Type:     clawv1alpha1.CredentialTypeGCP,
-				Provider: "anthropic",
-				GCP:      &clawv1alpha1.GCPConfig{Project: "p", Location: "us-east5"},
-			},
-			wantDomain: ".googleapis.com",
-		},
-		{
-			name: "explicit domain preserved",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "gemini",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "google",
-				Domain:   "custom-proxy.internal",
-			},
-			wantDomain: "custom-proxy.internal",
-			wantHeader: "x-goog-api-key",
-		},
-		{
-			name: "explicit apiKey preserved",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "gemini",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "google",
-				APIKey:   &clawv1alpha1.APIKeyConfig{Header: "x-custom-key"},
-			},
-			wantDomain: "generativelanguage.googleapis.com",
-			wantHeader: "x-custom-key",
-		},
-		{
-			name: "unknown provider with domain succeeds",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "custom",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "custom-llm",
-				Domain:   "api.custom-llm.com",
-				APIKey:   &clawv1alpha1.APIKeyConfig{Header: "x-api-key"},
-			},
-			wantDomain: "api.custom-llm.com",
-			wantHeader: "x-api-key",
-		},
-		{
-			name: "unknown provider without domain errors",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "custom",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "custom-llm",
-				APIKey:   &clawv1alpha1.APIKeyConfig{Header: "x-api-key"},
-			},
-			wantErr: "domain is required",
-		},
-		{
-			name: "unknown provider without apiKey errors",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "custom",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "custom-llm",
-				Domain:   "api.custom-llm.com",
-			},
-			wantErr: "apiKey config is required",
-		},
-		{
-			name: "no provider with domain and apiKey succeeds",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:   "legacy",
-				Type:   clawv1alpha1.CredentialTypeAPIKey,
-				Domain: "api.example.com",
-				APIKey: &clawv1alpha1.APIKeyConfig{Header: "x-token"},
-			},
-			wantDomain: "api.example.com",
-			wantHeader: "x-token",
-		},
-		{
-			name: "bearer type with no domain errors",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "custom",
-				Type:     clawv1alpha1.CredentialTypeBearer,
-				Provider: "custom-llm",
-			},
-			wantErr: "domain is required",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			cred := tt.cred
-			err := resolveProviderDefaults(&cred)
-			if tt.wantErr != "" {
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), tt.wantErr)
-				return
-			}
-			require.NoError(t, err)
-			assert.Equal(t, tt.wantDomain, cred.Domain)
-			if tt.wantHeader != "" {
-				require.NotNil(t, cred.APIKey)
-				assert.Equal(t, tt.wantHeader, cred.APIKey.Header)
-			}
-		})
-	}
-}
-
 func TestResolveProviderInfo(t *testing.T) {
 	tests := []struct {
 		name         string
@@ -952,402 +809,6 @@ func TestInjectProvidersIntoConfigMap(t *testing.T) {
 
 		primary := defaults["model"].(map[string]any)
 		assert.Equal(t, "anthropic/claude-haiku", primary["primary"], "fallback should be lexicographically first available model")
-	})
-}
-
-func TestUsesVertexSDK(t *testing.T) {
-	tests := []struct {
-		name string
-		cred clawv1alpha1.CredentialSpec
-		want bool
-	}{
-		{
-			name: "GCP + anthropic uses Vertex SDK",
-			cred: clawv1alpha1.CredentialSpec{Type: clawv1alpha1.CredentialTypeGCP, Provider: "anthropic"},
-			want: true,
-		},
-		{
-			name: "GCP + google does not use Vertex SDK",
-			cred: clawv1alpha1.CredentialSpec{Type: clawv1alpha1.CredentialTypeGCP, Provider: "google"},
-			want: false,
-		},
-		{
-			name: "GCP without provider does not use Vertex SDK",
-			cred: clawv1alpha1.CredentialSpec{Type: clawv1alpha1.CredentialTypeGCP},
-			want: false,
-		},
-		{
-			name: "apiKey + anthropic does not use Vertex SDK",
-			cred: clawv1alpha1.CredentialSpec{Type: clawv1alpha1.CredentialTypeAPIKey, Provider: "anthropic"},
-			want: false,
-		},
-		{
-			name: "GCP + meta uses Vertex SDK",
-			cred: clawv1alpha1.CredentialSpec{Type: clawv1alpha1.CredentialTypeGCP, Provider: "meta"},
-			want: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.want, usesVertexSDK(tt.cred))
-		})
-	}
-}
-
-func TestGenerateProxyConfigVertexSDK(t *testing.T) {
-	t.Run("should not create gateway route for GCP anthropic credential", func(t *testing.T) {
-		credentials := []clawv1alpha1.CredentialSpec{
-			{
-				Name:     "anthropic-vertex",
-				Type:     clawv1alpha1.CredentialTypeGCP,
-				Provider: "anthropic",
-				SecretRef: &clawv1alpha1.SecretRef{
-					Name: "vertex-sa",
-					Key:  "sa.json",
-				},
-				Domain: ".googleapis.com",
-				GCP: &clawv1alpha1.GCPConfig{
-					Project:  "my-project",
-					Location: "us-east5",
-				},
-			},
-		}
-
-		data, err := generateProxyConfig(credentials)
-		require.NoError(t, err)
-
-		var cfg proxyConfig
-		require.NoError(t, json.Unmarshal(data, &cfg))
-		require.Len(t, cfg.Routes, 1)
-		assert.Equal(t, "gcp", cfg.Routes[0].Injector)
-		assert.Empty(t, cfg.Routes[0].PathPrefix, "Vertex SDK provider should not have gateway path prefix")
-		assert.Empty(t, cfg.Routes[0].Upstream, "Vertex SDK provider should not have gateway upstream")
-	})
-
-	t.Run("should create gateway route for GCP google but not GCP anthropic", func(t *testing.T) {
-		credentials := []clawv1alpha1.CredentialSpec{
-			{
-				Name:     "gemini-vertex",
-				Type:     clawv1alpha1.CredentialTypeGCP,
-				Provider: "google",
-				SecretRef: &clawv1alpha1.SecretRef{
-					Name: "gcp-secret",
-					Key:  "sa.json",
-				},
-				Domain: ".googleapis.com",
-				GCP: &clawv1alpha1.GCPConfig{
-					Project:  "my-project",
-					Location: "us-central1",
-				},
-			},
-			{
-				Name:     "anthropic-vertex",
-				Type:     clawv1alpha1.CredentialTypeGCP,
-				Provider: "anthropic",
-				SecretRef: &clawv1alpha1.SecretRef{
-					Name: "vertex-sa",
-					Key:  "sa.json",
-				},
-				Domain: ".googleapis.com",
-				GCP: &clawv1alpha1.GCPConfig{
-					Project:  "my-project",
-					Location: "us-east5",
-				},
-			},
-		}
-
-		data, err := generateProxyConfig(credentials)
-		require.NoError(t, err)
-
-		var cfg proxyConfig
-		require.NoError(t, json.Unmarshal(data, &cfg))
-		require.Len(t, cfg.Routes, 2)
-
-		var googleRoute, anthropicRoute *proxyRoute
-		for i := range cfg.Routes {
-			if cfg.Routes[i].SAFilePath == "/etc/proxy/credentials/gemini-vertex/sa-key.json" {
-				googleRoute = &cfg.Routes[i]
-			}
-			if cfg.Routes[i].SAFilePath == "/etc/proxy/credentials/anthropic-vertex/sa-key.json" {
-				anthropicRoute = &cfg.Routes[i]
-			}
-		}
-
-		require.NotNil(t, googleRoute, "google GCP route should exist")
-		assert.Equal(t, "/gemini-vertex", googleRoute.PathPrefix, "google GCP should have gateway prefix")
-		assert.NotEmpty(t, googleRoute.Upstream, "google GCP should have gateway upstream")
-
-		require.NotNil(t, anthropicRoute, "anthropic GCP route should exist")
-		assert.Empty(t, anthropicRoute.PathPrefix, "anthropic GCP should not have gateway prefix")
-		assert.Empty(t, anthropicRoute.Upstream, "anthropic GCP should not have gateway upstream")
-	})
-}
-
-func TestInjectProvidersVertexSDK(t *testing.T) {
-	makeConfigMap := func(jsonContent string) []*unstructured.Unstructured {
-		cm := &unstructured.Unstructured{}
-		cm.SetKind(ConfigMapKind)
-		cm.SetName(ClawConfigMapName)
-		cm.Object["data"] = map[string]any{
-			"openclaw.json": jsonContent,
-		}
-		return []*unstructured.Unstructured{cm}
-	}
-
-	getConfig := func(t *testing.T, objects []*unstructured.Unstructured) map[string]any {
-		t.Helper()
-		raw, _, err := unstructured.NestedString(objects[0].Object, "data", "openclaw.json")
-		require.NoError(t, err)
-		var config map[string]any
-		require.NoError(t, json.Unmarshal([]byte(raw), &config))
-		return config
-	}
-
-	getProviders := func(t *testing.T, config map[string]any) map[string]any {
-		t.Helper()
-		models := config["models"].(map[string]any)
-		return models["providers"].(map[string]any)
-	}
-
-	t.Run("should map GCP anthropic to anthropic-vertex provider", func(t *testing.T) {
-		objects := makeConfigMap(`{"models":{"providers":{}}}`)
-		credentials := []clawv1alpha1.CredentialSpec{
-			{
-				Name:     "anthropic-vertex",
-				Type:     clawv1alpha1.CredentialTypeGCP,
-				Provider: "anthropic",
-				Domain:   ".googleapis.com",
-				GCP: &clawv1alpha1.GCPConfig{
-					Project:  "my-project",
-					Location: "us-east5",
-				},
-			},
-		}
-
-		require.NoError(t, injectProvidersIntoConfigMap(objects, credentials))
-
-		config := getConfig(t, objects)
-		providers := getProviders(t, config)
-
-		assert.NotContains(t, providers, "anthropic", "should not have base anthropic provider")
-		require.Contains(t, providers, "anthropic-vertex")
-
-		av := providers["anthropic-vertex"].(map[string]any)
-		assert.Equal(t, "https://us-east5-aiplatform.googleapis.com", av["baseUrl"])
-		assert.Equal(t, "gcp-vertex-credentials", av["apiKey"])
-		assert.Equal(t, "anthropic-messages", av["api"])
-	})
-
-	t.Run("should remap model entries from anthropic to anthropic-vertex", func(t *testing.T) {
-		configJSON := `{
-			"models": {"providers": {}},
-			"agents": {
-				"defaults": {
-					"model": {"primary": "anthropic/claude-sonnet-4-6"},
-					"models": {
-						"anthropic/claude-sonnet-4-6": {"alias": "Claude Sonnet"},
-						"anthropic/claude-haiku-4-5": {"alias": "Claude Haiku"},
-						"google/gemini-flash": {"alias": "Gemini Flash"}
-					}
-				}
-			}
-		}`
-		objects := makeConfigMap(configJSON)
-		credentials := []clawv1alpha1.CredentialSpec{
-			{
-				Name:     "gemini",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "google",
-				Domain:   "generativelanguage.googleapis.com",
-			},
-			{
-				Name:     "anthropic-vertex",
-				Type:     clawv1alpha1.CredentialTypeGCP,
-				Provider: "anthropic",
-				Domain:   ".googleapis.com",
-				GCP: &clawv1alpha1.GCPConfig{
-					Project:  "my-project",
-					Location: "us-east5",
-				},
-			},
-		}
-
-		require.NoError(t, injectProvidersIntoConfigMap(objects, credentials))
-
-		config := getConfig(t, objects)
-		agents := config["agents"].(map[string]any)
-		defaults := agents["defaults"].(map[string]any)
-		modelAliases := defaults["models"].(map[string]any)
-
-		assert.NotContains(t, modelAliases, "anthropic/claude-sonnet-4-6", "original anthropic key should be removed")
-		assert.NotContains(t, modelAliases, "anthropic/claude-haiku-4-5", "original anthropic key should be removed")
-		assert.Contains(t, modelAliases, "anthropic-vertex/claude-sonnet-4-6", "remapped key should exist")
-		assert.Contains(t, modelAliases, "anthropic-vertex/claude-haiku-4-5", "remapped key should exist")
-		assert.Contains(t, modelAliases, "google/gemini-flash", "google key should be preserved")
-
-		primary := defaults["model"].(map[string]any)
-		assert.Equal(t, "anthropic-vertex/claude-sonnet-4-6", primary["primary"], "primary should be remapped")
-	})
-
-	t.Run("should not remap when both base and vertex providers exist", func(t *testing.T) {
-		configJSON := `{
-			"models": {"providers": {}},
-			"agents": {
-				"defaults": {
-					"model": {"primary": "anthropic/direct-model"},
-					"models": {
-						"anthropic/direct-model": {"alias": "Direct"}
-					}
-				}
-			}
-		}`
-		objects := makeConfigMap(configJSON)
-		credentials := []clawv1alpha1.CredentialSpec{
-			{
-				Name:     "claude-direct",
-				Type:     clawv1alpha1.CredentialTypeBearer,
-				Provider: "anthropic",
-				Domain:   "api.anthropic.com",
-			},
-			{
-				Name:     "claude-vertex",
-				Type:     clawv1alpha1.CredentialTypeGCP,
-				Provider: "anthropic",
-				Domain:   ".googleapis.com",
-				GCP: &clawv1alpha1.GCPConfig{
-					Project:  "my-project",
-					Location: "us-east5",
-				},
-			},
-		}
-
-		require.NoError(t, injectProvidersIntoConfigMap(objects, credentials))
-
-		config := getConfig(t, objects)
-		agents := config["agents"].(map[string]any)
-		defaults := agents["defaults"].(map[string]any)
-		modelAliases := defaults["models"].(map[string]any)
-
-		assert.Contains(t, modelAliases, "anthropic/direct-model", "should keep base provider models when both exist")
-
-		primary := defaults["model"].(map[string]any)
-		assert.Equal(t, "anthropic/direct-model", primary["primary"], "primary should not be remapped")
-	})
-
-	t.Run("should reject duplicate vertex providers", func(t *testing.T) {
-		objects := makeConfigMap(`{"models":{"providers":{}}}`)
-		credentials := []clawv1alpha1.CredentialSpec{
-			{
-				Name: "claude-vertex-1", Type: clawv1alpha1.CredentialTypeGCP, Provider: "anthropic",
-				Domain: ".googleapis.com", GCP: &clawv1alpha1.GCPConfig{Project: "p1", Location: "us-east5"},
-			},
-			{
-				Name: "claude-vertex-2", Type: clawv1alpha1.CredentialTypeGCP, Provider: "anthropic",
-				Domain: ".googleapis.com", GCP: &clawv1alpha1.GCPConfig{Project: "p2", Location: "us-east5"},
-			},
-		}
-
-		err := injectProvidersIntoConfigMap(objects, credentials)
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "duplicate provider")
-		assert.Contains(t, err.Error(), "anthropic-vertex")
-	})
-}
-
-func TestConfigureClawDeploymentForVertex(t *testing.T) {
-	makeDeployment := func() []*unstructured.Unstructured {
-		dep := &unstructured.Unstructured{}
-		dep.SetKind(DeploymentKind)
-		dep.SetName(ClawDeploymentName)
-		dep.Object["spec"] = map[string]any{
-			"template": map[string]any{
-				"spec": map[string]any{
-					"containers": []any{
-						map[string]any{
-							"name": ClawGatewayContainerName,
-							"env": []any{
-								map[string]any{"name": "HOME", "value": "/home/node"},
-							},
-							"volumeMounts": []any{},
-						},
-					},
-					"volumes": []any{},
-				},
-			},
-		}
-		return []*unstructured.Unstructured{dep}
-	}
-
-	t.Run("should add vertex env vars and volume mount", func(t *testing.T) {
-		objects := makeDeployment()
-		credentials := []clawv1alpha1.CredentialSpec{
-			{
-				Name:     "anthropic-vertex",
-				Type:     clawv1alpha1.CredentialTypeGCP,
-				Provider: "anthropic",
-				Domain:   ".googleapis.com",
-				GCP: &clawv1alpha1.GCPConfig{
-					Project:  "my-project",
-					Location: "us-east5",
-				},
-			},
-		}
-
-		require.NoError(t, configureClawDeploymentForVertex(objects, credentials))
-
-		containers, _, _ := unstructured.NestedSlice(objects[0].Object, "spec", "template", "spec", "containers")
-		container := containers[0].(map[string]any)
-		envVars := container["env"].([]any)
-
-		var adcEnv, projectEnv map[string]any
-		for _, e := range envVars {
-			env := e.(map[string]any)
-			switch env["name"] {
-			case "GOOGLE_APPLICATION_CREDENTIALS":
-				adcEnv = env
-			case "ANTHROPIC_VERTEX_PROJECT_ID":
-				projectEnv = env
-			}
-		}
-
-		require.NotNil(t, adcEnv, "GOOGLE_APPLICATION_CREDENTIALS should be set")
-		assert.Equal(t, "/etc/vertex-adc/adc.json", adcEnv["value"])
-
-		require.NotNil(t, projectEnv, "ANTHROPIC_VERTEX_PROJECT_ID should be set")
-		assert.Equal(t, "my-project", projectEnv["value"])
-
-		volumeMounts := container["volumeMounts"].([]any)
-		require.Len(t, volumeMounts, 1)
-		vm := volumeMounts[0].(map[string]any)
-		assert.Equal(t, "vertex-adc", vm["name"])
-		assert.Equal(t, "/etc/vertex-adc", vm["mountPath"])
-		assert.Equal(t, true, vm["readOnly"])
-
-		volumes, _, _ := unstructured.NestedSlice(objects[0].Object, "spec", "template", "spec", "volumes")
-		require.Len(t, volumes, 1)
-		vol := volumes[0].(map[string]any)
-		assert.Equal(t, "vertex-adc", vol["name"])
-		cmRef := vol["configMap"].(map[string]any)
-		assert.Equal(t, ClawVertexADCConfigMapName, cmRef["name"])
-	})
-
-	t.Run("should be no-op when no vertex credentials exist", func(t *testing.T) {
-		objects := makeDeployment()
-		credentials := []clawv1alpha1.CredentialSpec{
-			{
-				Name:     "gemini",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "google",
-				Domain:   "generativelanguage.googleapis.com",
-			},
-		}
-
-		require.NoError(t, configureClawDeploymentForVertex(objects, credentials))
-
-		containers, _, _ := unstructured.NestedSlice(objects[0].Object, "spec", "template", "spec", "containers")
-		container := containers[0].(map[string]any)
-		envVars := container["env"].([]any)
-		assert.Len(t, envVars, 1, "should only have original HOME env var")
 	})
 }
